A few fixes for C-periodic C++ sim, but it is fundamentally wrong
There is no obvious way to construct 100% flippable clusters given
the weights associated with the non-symmetric links at the boundaries.
Here, there are 4 distinct Boltzmann weights instead of just two,
because flipping through some height h is NOT a symmetry. The only
cluster update that would work is flips through h=0, but this is not
ergodic!